change proposal card limit, restructure minionActionFnNames to minionUtils
