+++ src/utils/proposalUtils.jsx
@@ -59,13 +59,6 @@ export const MINION_TYPES = {
   UBER: 'UberHaus minion',
 };
 
-export const MINION_ACTION_FUNCTION_NAMES = {
-  VANILLA_MINION: 'actions',
-  SAFE_MINION: 'actions',
-  UBERHAUS_MINION: 'appointments',
-  SUPERFLUID_MINION: 'streams',
-};
-
 export const inQueue = proposal => {
   const now = new Date() / 1000 || 0;
   return now < +proposal.votingPeriodStarts;
